add: EditProfile 파일 추가(리팩토링)
